Fixes to remove compiler warnings
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@1707 610dfd11-2700-0410-9b68-9ae28eac5983

+++ src/Anchor.cpp
@@ -111,7 +111,7 @@ void Anchor::GeometrySelected::update() {
 
         bool file_exist = true;
         AutoRef< Node > new_world;
-        unsigned int vp_list_size =
+        size_t vp_list_size =
           X3DViewpointNode::getAllViewpoints().size();
         int prev_outputlevel = Console.getOutputLevel();
         Console.setOutputLevel( 100 );
